[FEATURE] Keyboard autocomplete (by zakm123): merged and some cleanup
